+++ src/algos/diffie_hellmann.rs
@@ -9,7 +9,7 @@ impl DiffieHellman {
 }
 
 impl Encryptor for DiffieHellman {
-    fn encrypt(&self, message: &str, shared_key: &u32) -> Result<String, String> {
+    fn encrypt(&self, message: &str, shared_key: &u16) -> Result<String, String> {
         let encrypted_msg = caesar_cipher(message, *shared_key);
 
         Ok(encrypted_msg)
@@ -17,7 +17,7 @@ impl Encryptor for DiffieHellman {
 }
 
 impl Decryptor for DiffieHellman {
-    fn decrypt(&self, message: &str, shared_key: &u32) -> Result<String, String> {
+    fn decrypt(&self, message: &str, shared_key: &u16) -> Result<String, String> {
         let decrypted_msg = caesar_cipher(message, 26 - (shared_key % 26));
 
         Ok(decrypted_msg)
